Routing change for firebase instances
